Add MessageAction
Previously, it was quite hard to add a non-actionable message
to the output of a Batch. One basically had to add a Process
which just ran `echo My Message`.

Since Batches have a Symfony OutputInterface readily available
that should have been the default output-only option from the
beginning which now is achieved by `Batch::addMessage` and
- in effect - the MessageAction.